requirements.txt
langchain-anthropic>=0.0.9
langchain-core>=0.1.17
langgraph>=0.0.11
langchain-google-community>=0.0.7
boto3>=1.34.0
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
gradio>=4.19.2 